+++ src/core/value.h
@@ -69,8 +69,8 @@ typedef struct Obj Obj;
 typedef struct ObjString ObjString;
 
 typedef struct {
-    int capacity;
-    int count;
+    size_t capacity;
+    size_t count;
     Value* values;
 } ValueArray;
